+++ themes-default/slim/src/components/index.js
@@ -10,6 +10,7 @@ export { default as Config } from './config.vue';
 export { default as ConfigAnime } from './config-anime.vue';
 export { default as ConfigGeneral } from './config-general.vue';
 export { default as ConfigPostProcessing } from './config-post-processing.vue';
+export { default as ConfigProviders } from './config-providers.vue';
 export { default as ConfigNotifications } from './config-notifications.vue';
 export { default as ConfigSearch } from './config-search.vue';
 export { default as DisplayShow } from './display-show.vue';