+++ themes-default/slim/src/components/snatch-selection.vue
@@ -133,5 +133,8 @@ export default {
 };
 </script>
 
-<style scoped src='../style/vgt-table.css'>
+<style scoped>
+.show-history {
+    margin-bottom: 10px;
+}
 </style>