+++ themes-default/slim/src/store/modules/config/general.js
@@ -12,6 +12,7 @@ const state = {
     subtitles: {
         enabled: null
     },
+    brokenProviders: [],
     logs: {
         debug: null,
         dbDebug: null,
@@ -106,7 +107,13 @@ const state = {
         period: null
     },
     // Remove themeName when we get fully rid of MEDUSA.config.
-    themeName: null
+    themeName: null,
+    providers: {
+        prowlarr: {
+            url: null,
+            apikey: null
+        }
+    }
 };
 
 const mutations = {