+++ themes-default/slim/webpack.config.js
@@ -210,6 +210,17 @@ const webpackConfig = (env, mode) => ({
                 ]
             },
             {
+                test: /\.s[ac]ss$/i,
+                use: [
+                    // Creates `style` nodes from JS strings
+                    'style-loader',
+                    // Translates CSS into CommonJS
+                    'css-loader',
+                    // Compiles Sass to CSS
+                    'sass-loader'
+                ]
+            },
+            {
                 test: /\.(woff2?|ttf|eot|svg)$/,
                 loader: 'file-loader',
                 options: {